+++ Documentation/CHANGELOG.md
@@ -29,6 +29,7 @@ interpolation for problems with CRRA utility. See [#888](https://github.com/econ
 * Corrects location of constructor documentation to class string for Sphinx rendering [#908](https://github.com/econ-ark/HARK/pull/908)
 * Adds a module with tools for parsing and using various income calibrations from the literature. It includes the option of using life-cycle profiles of income shock variances from [Sabelhaus and Song (2010)](https://www.sciencedirect.com/science/article/abs/pii/S0304393210000358). See [#921](https://github.com/econ-ark/HARK/pull/921), [#941](https://github.com/econ-ark/HARK/pull/941).
 * remove "Now" from model variable names [#936](https://github.com/econ-ark/HARK/pull/936)
+* remove Model.__call__; use Model init in Market and AgentType init to standardize on parameters dictionary [#947](https://github.com/econ-ark/HARK/issues/947)
 * Moves state MrkvNow to shocks['Mrkv'] in AggShockMarkov and KrusellSmith models [#935](https://github.com/econ-ark/HARK/pull/935)
 * Replaces `ConsIndShock`'s `init_lifecycle` with an actual life-cycle calibration [#951](https://github.com/econ-ark/HARK/pull/951).
 
@@ -40,6 +41,8 @@ interpolation for problems with CRRA utility. See [#888](https://github.com/econ
 * Rename IncomeDstn to IncShkDstn
 * AgentType simulate() method now returns history. [#916](https://github.com/econ-ark/HARK/pull/916)
 * Rename DiscreteDistribution.drawDiscrete() to draw()
+* Update documentation and warnings around IncShkDstn [#955](https://github.com/econ-ark/HARK/pull/955)
+* Adds csv files to `MANIFEST.in`. [957](https://github.com/econ-ark/HARK/pull/957)
 
 ### 0.10.8